[AI Assistant] Move alert details context from APM api to Obs api
