+++ packages/kbn-apm-synthtrace/index.ts
@@ -9,7 +9,8 @@
 export { createLogger, LogLevel } from './src/lib/utils/create_logger';
 
 export { ApmSynthtraceEsClient } from './src/lib/apm/client/apm_synthtrace_es_client';
-export { ApmSynthtraceKibanaClient } from './src/lib/apm/client/apm_synthtrace_kibana_client';
+export { SynthtraceKibanaClient } from './src/lib/apm/client/apm_synthtrace_kibana_client';
+
 export { InfraSynthtraceEsClient } from './src/lib/infra/infra_synthtrace_es_client';
 export { InfraSynthtraceKibanaClient } from './src/lib/infra/infra_synthtrace_kibana_client';
 export { MonitoringSynthtraceEsClient } from './src/lib/monitoring/monitoring_synthtrace_es_client';
