+++ packages/kbn-apm-synthtrace/src/cli/utils/get_kibana_client.ts
@@ -6,11 +6,11 @@
  * Side Public License, v 1.
  */
 
-import { ApmSynthtraceKibanaClient } from '../../lib/apm/client/apm_synthtrace_kibana_client';
+import { SynthtraceKibanaClient } from '../../lib/apm/client/apm_synthtrace_kibana_client';
 import { Logger } from '../../lib/utils/create_logger';
 
 export function getKibanaClient({ target, logger }: { target: string; logger: Logger }) {
-  const kibanaClient = new ApmSynthtraceKibanaClient({
+  const kibanaClient = new SynthtraceKibanaClient({
     logger,
     target,
   });
