+++ x-pack/plugins/observability_solution/apm/ftr_e2e/cypress_test_runner.ts
@@ -5,7 +5,7 @@
  * 2.0.
  */
 
-import { ApmSynthtraceKibanaClient, createLogger, LogLevel } from '@kbn/apm-synthtrace';
+import { SynthtraceKibanaClient, createLogger, LogLevel } from '@kbn/apm-synthtrace';
 import cypress from 'cypress';
 import path from 'path';
 import Url from 'url';
@@ -39,7 +39,7 @@ export async function cypressTestRunner({ getService }: FtrProviderContext) {
   });
 
   const esRequestTimeout = config.get('timeouts.esRequestTimeout');
-  const kibanaClient = new ApmSynthtraceKibanaClient({
+  const kibanaClient = new SynthtraceKibanaClient({
     logger: createLogger(LogLevel.info),
     target: kibanaUrl,
   });
