+++ x-pack/plugins/observability_solution/apm/server/routes/assistant_functions/get_observability_alert_details_context/index.ts
@@ -5,17 +5,16 @@
  * 2.0.
  */
 
-import type { ScopedAnnotationsClient } from '@kbn/observability-plugin/server';
-import type { ElasticsearchClient } from '@kbn/core-elasticsearch-server';
-import type { CoreRequestHandlerContext, Logger } from '@kbn/core/server';
+import { Logger } from '@kbn/core/server';
 import {
-  AlertDetailsContextualInsight,
   AlertDetailsContextualInsightsHandlerQuery,
+  AlertDetailsContextualInsightsRequestContext,
 } from '@kbn/observability-plugin/server/services';
 import moment from 'moment';
-import type { MlClient } from '../../../lib/helpers/get_ml_client';
-import type { APMEventClient } from '../../../lib/helpers/create_es_client/create_apm_event_client';
-import type { ApmAlertsClient } from '../../../lib/helpers/get_apm_alerts_client';
+import { getApmAlertsClient } from '../../../lib/helpers/get_apm_alerts_client';
+import { getApmEventClient } from '../../../lib/helpers/get_apm_event_client';
+import { getMlClient } from '../../../lib/helpers/get_ml_client';
+import { getRandomSampler } from '../../../lib/helpers/get_random_sampler';
 import { getApmServiceSummary } from '../get_apm_service_summary';
 import { getAssistantDownstreamDependencies } from '../get_apm_downstream_dependencies';
 import { getLogCategories } from '../get_log_categories';
@@ -24,157 +23,196 @@ import { getServiceNameFromSignals } from './get_service_name_from_signals';
 import { getContainerIdFromSignals } from './get_container_id_from_signals';
 import { getApmAlertDetailsContextPrompt } from './get_apm_alert_details_context_prompt';
 import { getExitSpanChangePoints, getServiceChangePoints } from '../get_changepoints';
+import { APMRouteHandlerResources } from '../../apm_routes/register_apm_server_routes';
 
-export async function getObservabilityAlertDetailsContext({
-  coreContext,
-  annotationsClient,
-  apmAlertsClient,
-  apmEventClient,
-  esClient,
-  logger,
-  mlClient,
-  query,
-}: {
-  coreContext: Pick<CoreRequestHandlerContext, 'elasticsearch' | 'uiSettings'>;
-  annotationsClient?: ScopedAnnotationsClient;
-  apmAlertsClient: ApmAlertsClient;
-  apmEventClient: APMEventClient;
-  esClient: ElasticsearchClient;
-  logger: Logger;
-  mlClient?: MlClient;
-  query: AlertDetailsContextualInsightsHandlerQuery;
-}): Promise<AlertDetailsContextualInsight[]> {
-  const alertStartedAt = query.alert_started_at;
-  const serviceEnvironment = query['service.environment'];
-  const hostName = query['host.name'];
-  const kubernetesPodName = query['kubernetes.pod.name'];
-  const [serviceName, containerId] = await Promise.all([
-    getServiceNameFromSignals({
-      query,
-      esClient,
-      coreContext,
-      apmEventClient,
-    }),
-    getContainerIdFromSignals({
-      query,
-      esClient,
-      coreContext,
-      apmEventClient,
-    }),
-  ]);
-
-  async function handleError<T>(cb: () => Promise<T>): Promise<T | undefined> {
-    try {
-      return await cb();
-    } catch (error) {
-      logger.error('Error while fetching observability alert details context');
-      logger.error(error);
-      return;
-    }
-  }
-
-  const serviceSummaryPromise = serviceName
-    ? handleError(() =>
-        getApmServiceSummary({
-          apmEventClient,
-          annotationsClient,
-          esClient,
-          apmAlertsClient,
-          mlClient,
-          logger,
-          arguments: {
-            'service.name': serviceName,
-            'service.environment': serviceEnvironment,
-            start: moment(alertStartedAt).subtract(5, 'minute').toISOString(),
-            end: alertStartedAt,
-          },
-        })
-      )
-    : undefined;
-
-  const downstreamDependenciesPromise = serviceName
-    ? handleError(() =>
-        getAssistantDownstreamDependencies({
-          apmEventClient,
-          arguments: {
-            'service.name': serviceName,
-            'service.environment': serviceEnvironment,
-            start: moment(alertStartedAt).subtract(5, 'minute').toISOString(),
-            end: alertStartedAt,
-          },
-        })
-      )
-    : undefined;
-
-  const logCategoriesPromise = handleError(() =>
-    getLogCategories({
-      esClient,
-      coreContext,
-      arguments: {
-        start: moment(alertStartedAt).subtract(5, 'minute').toISOString(),
-        end: alertStartedAt,
-        'service.name': serviceName,
-        'host.name': hostName,
-        'container.id': containerId,
-        'kubernetes.pod.name': kubernetesPodName,
+export const getAlertDetailsContextHandler = (
+  resourcePlugins: APMRouteHandlerResources['plugins'],
+  logger: Logger
+) => {
+  return async (
+    requestContext: AlertDetailsContextualInsightsRequestContext,
+    query: AlertDetailsContextualInsightsHandlerQuery
+  ) => {
+    const resources = {
+      getApmIndices: async () => {
+        const coreContext = await requestContext.core;
+        return resourcePlugins.apmDataAccess.setup.getApmIndices(coreContext.savedObjects.client);
       },
-    })
-  );
+      request: requestContext.request,
+      params: { query: { _inspect: false } },
+      plugins: resourcePlugins,
+      context: {
+        core: requestContext.core,
+        licensing: requestContext.licensing,
+        alerting: resourcePlugins.alerting!.start().then((startContract) => {
+          return {
+            getRulesClient() {
+              return startContract.getRulesClientWithRequest(requestContext.request);
+            },
+          };
+        }),
+        rac: resourcePlugins.ruleRegistry.start().then((startContract) => {
+          return {
+            getAlertsClient() {
+              return startContract.getRacClientWithRequest(requestContext.request);
+            },
+          };
+        }),
+      },
+    };
 
-  const serviceChangePointsPromise = handleError(() =>
-    getServiceChangePoints({
-      apmEventClient,
-      alertStartedAt,
+    const [apmEventClient, annotationsClient, apmAlertsClient, coreContext, mlClient] =
+      await Promise.all([
+        getApmEventClient(resources),
+        resourcePlugins.observability.setup.getScopedAnnotationsClient(
+          resources.context,
+          requestContext.request
+        ),
+        getApmAlertsClient(resources),
+        requestContext.core,
+        getMlClient(resources),
+        getRandomSampler({
+          security: resourcePlugins.security,
+          probability: 1,
+          request: requestContext.request,
+        }),
+      ]);
+    const esClient = coreContext.elasticsearch.client.asCurrentUser;
+
+    const alertStartedAt = query.alert_started_at;
+    const serviceEnvironment = query['service.environment'];
+    const hostName = query['host.name'];
+    const kubernetesPodName = query['kubernetes.pod.name'];
+    const [serviceName, containerId] = await Promise.all([
+      getServiceNameFromSignals({
+        query,
+        esClient,
+        coreContext,
+        apmEventClient,
+      }),
+      getContainerIdFromSignals({
+        query,
+        esClient,
+        coreContext,
+        apmEventClient,
+      }),
+    ]);
+
+    async function handleError<T>(cb: () => Promise<T>): Promise<T | undefined> {
+      try {
+        return await cb();
+      } catch (error) {
+        logger.error('Error while fetching observability alert details context');
+        logger.error(error);
+        return;
+      }
+    }
+
+    const serviceSummaryPromise = serviceName
+      ? handleError(() =>
+          getApmServiceSummary({
+            apmEventClient,
+            annotationsClient,
+            esClient,
+            apmAlertsClient,
+            mlClient,
+            logger,
+            arguments: {
+              'service.name': serviceName,
+              'service.environment': serviceEnvironment,
+              start: moment(alertStartedAt).subtract(5, 'minute').toISOString(),
+              end: alertStartedAt,
+            },
+          })
+        )
+      : undefined;
+
+    const downstreamDependenciesPromise = serviceName
+      ? handleError(() =>
+          getAssistantDownstreamDependencies({
+            apmEventClient,
+            arguments: {
+              'service.name': serviceName,
+              'service.environment': serviceEnvironment,
+              start: moment(alertStartedAt).subtract(15, 'minute').toISOString(),
+              end: alertStartedAt,
+            },
+          })
+        )
+      : undefined;
+
+    const logCategoriesPromise = handleError(() =>
+      getLogCategories({
+        esClient,
+        coreContext,
+        arguments: {
+          start: moment(alertStartedAt).subtract(15, 'minute').toISOString(),
+          end: alertStartedAt,
+          'service.name': serviceName,
+          'host.name': hostName,
+          'container.id': containerId,
+          'kubernetes.pod.name': kubernetesPodName,
+        },
+      })
+    );
+
+    const serviceChangePointsPromise = handleError(() =>
+      getServiceChangePoints({
+        apmEventClient,
+        start: moment(alertStartedAt).subtract(6, 'hours').toISOString(),
+        end: alertStartedAt,
+        serviceName,
+        serviceEnvironment,
+        transactionType: query['transaction.type'],
+        transactionName: query['transaction.name'],
+      })
+    );
+
+    const exitSpanChangePointsPromise = handleError(() =>
+      getExitSpanChangePoints({
+        apmEventClient,
+        start: moment(alertStartedAt).subtract(6, 'hours').toISOString(),
+        end: alertStartedAt,
+        serviceName,
+        serviceEnvironment,
+      })
+    );
+
+    const anomaliesPromise = handleError(() =>
+      getAnomalies({
+        start: moment(alertStartedAt).subtract(1, 'hour').valueOf(),
+        end: moment(alertStartedAt).valueOf(),
+        environment: serviceEnvironment,
+        mlClient,
+        logger,
+      })
+    );
+
+    const [
+      serviceSummary,
+      downstreamDependencies,
+      logCategories,
+      serviceChangePoints,
+      exitSpanChangePoints,
+      anomalies,
+    ] = await Promise.all([
+      serviceSummaryPromise,
+      downstreamDependenciesPromise,
+      logCategoriesPromise,
+      serviceChangePointsPromise,
+      exitSpanChangePointsPromise,
+      anomaliesPromise,
+    ]);
+
+    return getApmAlertDetailsContextPrompt({
       serviceName,
       serviceEnvironment,
-      transactionType: query['transaction.type'],
-      transactionName: query['transaction.name'],
-    })
-  );
-
-  const exitSpanChangePointsPromise = handleError(() =>
-    getExitSpanChangePoints({
-      apmEventClient,
-      alertStartedAt,
-      serviceName,
-      serviceEnvironment,
-    })
-  );
-
-  const anomaliesPromise = handleError(() =>
-    getAnomalies({
-      start: moment(alertStartedAt).subtract(1, 'hour').valueOf(),
-      end: moment(alertStartedAt).valueOf(),
-      environment: serviceEnvironment,
-      mlClient,
-      logger,
-    })
-  );
-
-  const [
-    serviceSummary,
-    downstreamDependencies,
-    logCategories,
-    serviceChangePoints,
-    exitSpanChangePoints,
-    anomalies,
-  ] = await Promise.all([
-    serviceSummaryPromise,
-    downstreamDependenciesPromise,
-    logCategoriesPromise,
-    serviceChangePointsPromise,
-    exitSpanChangePointsPromise,
-    anomaliesPromise,
-  ]);
-
-  return getApmAlertDetailsContextPrompt({
-    serviceName,
-    serviceEnvironment,
-    serviceSummary,
-    downstreamDependencies,
-    logCategories,
-    serviceChangePoints,
-    exitSpanChangePoints,
-    anomalies,
-  });
-}
+      serviceSummary,
+      downstreamDependencies,
+      logCategories,
+      serviceChangePoints,
+      exitSpanChangePoints,
+      anomalies,
+    });
+  };
+};
