+++ x-pack/plugins/observability_solution/apm/server/routes/assistant_functions/route.ts
@@ -6,16 +6,8 @@
  */
 import * as t from 'io-ts';
 import { omit } from 'lodash';
-import {
-  AlertDetailsContextualInsight,
-  observabilityAlertDetailsContextRt,
-} from '@kbn/observability-plugin/server/services';
-import { getApmAlertsClient } from '../../lib/helpers/get_apm_alerts_client';
 import { getApmEventClient } from '../../lib/helpers/get_apm_event_client';
-import { getMlClient } from '../../lib/helpers/get_ml_client';
-import { getRandomSampler } from '../../lib/helpers/get_random_sampler';
 import { createApmServerRoute } from '../apm_routes/create_apm_server_route';
-import { getObservabilityAlertDetailsContext } from './get_observability_alert_details_context';
 
 import {
   downstreamDependenciesRouteRt,
@@ -23,49 +15,6 @@ import {
   type APMDownstreamDependency,
 } from './get_apm_downstream_dependencies';
 import { getApmTimeseries, getApmTimeseriesRt, type ApmTimeseries } from './get_apm_timeseries';
-
-const getObservabilityAlertDetailsContextRoute = createApmServerRoute({
-  endpoint: 'GET /internal/apm/assistant/alert_details_contextual_insights',
-  options: {
-    tags: ['access:apm'],
-  },
-
-  params: t.type({
-    query: observabilityAlertDetailsContextRt,
-  }),
-  handler: async (resources): Promise<{ context: AlertDetailsContextualInsight[] }> => {
-    const { context, request, plugins, logger, params } = resources;
-    const { query } = params;
-
-    const [apmEventClient, annotationsClient, coreContext, apmAlertsClient, mlClient] =
-      await Promise.all([
-        getApmEventClient(resources),
-        plugins.observability.setup.getScopedAnnotationsClient(context, request),
-        context.core,
-        getApmAlertsClient(resources),
-        getMlClient(resources),
-        getRandomSampler({
-          security: resources.plugins.security,
-          probability: 1,
-          request: resources.request,
-        }),
-      ]);
-    const esClient = coreContext.elasticsearch.client.asCurrentUser;
-
-    const obsAlertContext = await getObservabilityAlertDetailsContext({
-      coreContext,
-      annotationsClient,
-      apmAlertsClient,
-      apmEventClient,
-      esClient,
-      logger,
-      mlClient,
-      query,
-    });
-
-    return { context: obsAlertContext };
-  },
-});
 
 const getApmTimeSeriesRoute = createApmServerRoute({
   endpoint: 'POST /internal/apm/assistant/get_apm_timeseries',
@@ -120,6 +69,5 @@ const getDownstreamDependenciesRoute = createApmServerRoute({
 
 export const assistantRouteRepository = {
   ...getApmTimeSeriesRoute,
-  ...getObservabilityAlertDetailsContextRoute,
   ...getDownstreamDependenciesRoute,
 };
