+++ x-pack/plugins/observability_solution/observability/server/routes/types.ts
@@ -7,12 +7,15 @@
 import type { EndpointOf, ReturnOf, ServerRouteRepository } from '@kbn/server-route-repository';
 import { KibanaRequest, Logger } from '@kbn/core/server';
 
-import { ObservabilityServerRouteRepository } from './get_global_observability_server_route_repository';
+import {
+  ObservabilityServerRouteRepository,
+  APIEndpoint,
+} from './get_global_observability_server_route_repository';
 import { ObservabilityRequestHandlerContext } from '../types';
 import { RegisterRoutesDependencies } from './register_routes';
 import { ObservabilityConfig } from '..';
 
-export type { ObservabilityServerRouteRepository };
+export type { ObservabilityServerRouteRepository, APIEndpoint };
 
 export interface ObservabilityRouteHandlerResources {
   context: ObservabilityRequestHandlerContext;
