+++ x-pack/plugins/observability_solution/observability_ai_assistant_app/scripts/evaluation/setup_synthtrace.ts
@@ -9,7 +9,7 @@ import {
   ApmSynthtraceEsClient,
   InfraSynthtraceEsClient,
   LogsSynthtraceEsClient,
-  ApmSynthtraceKibanaClient,
+  SynthtraceKibanaClient,
 } from '@kbn/apm-synthtrace';
 import { ToolingLog } from '@kbn/tooling-log';
 import { isPromise } from 'util/types';
@@ -54,7 +54,7 @@ export async function setupSynthtrace({
       return result;
     },
   };
-  const kibanaClient = new ApmSynthtraceKibanaClient({
+  const kibanaClient = new SynthtraceKibanaClient({
     target,
     logger,
   });
