+++ x-pack/test/observability_api_integration/common/ftr_provider_context.ts
@@ -5,4 +5,12 @@
  * 2.0.
  */
 
+import { GenericFtrProviderContext } from '@kbn/test';
+import { services } from '../../api_integration/services';
+import { CustomApiTestServices } from './config';
+
 export type { FtrProviderContext } from '../../api_integration/ftr_provider_context';
+export type ObsFtrProviderContext = GenericFtrProviderContext<
+  typeof services & CustomApiTestServices,
+  {}
+>;
